refactor: AuthorizationExtractor component로 변경
