+++ src/main/java/nextstep/common/AuthenticationPrincipalArgumentResolver.java
@@ -2,7 +2,7 @@ package nextstep.common;
 
 import lombok.AllArgsConstructor;
 import nextstep.auth.JwtTokenProvider;
-import nextstep.auth.utils.AuthorizationExtractor;
+import nextstep.auth.AuthorizationExtractor;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
@@ -17,6 +17,7 @@ import javax.servlet.http.HttpServletRequest;
 public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
 
     private JwtTokenProvider jwtTokenProvider;
+    private AuthorizationExtractor authorizationExtractor;
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
@@ -25,7 +26,7 @@ public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArg
 
     @Override
     public String resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
-        String accessToken = AuthorizationExtractor.extract(webRequest.getNativeRequest(HttpServletRequest.class));
+        String accessToken = authorizationExtractor.extract(webRequest.getNativeRequest(HttpServletRequest.class));
         return jwtTokenProvider.getPrincipal(accessToken);
     }
 }
